修改 附件上传和删除的代码

+++ src/main/java/com/usedmarket/service/impl/AttachmentServiceImpl.java
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.List;
+
 /**
  * Created by huangMP on 2016/10/22.
  * decription : 用户信息 Service层实现
@@ -40,15 +42,7 @@ public class AttachmentServiceImpl implements AttachmentService {
         attachment.setFileType(fileType);
         attachment.setContentId(contentId);
 
-        //执行上传 返回真实文件名
-        String fileName = FileUpload.fileUp(file, fileAbsoluteathPath, UuidUtil.get32UUID());
-        //得到压缩图文件名
-        String narrowImageFileName = "_" + fileName;
-        //进行压缩
-        NarrowImage.imageNarrow(fileAbsoluteathPath, narrowImageFileName, fileName, 5);
-
-        attachment.setFilePath(fileRelativePath + fileName);
-        attachment.setNarrowImagePath(fileRelativePath + narrowImageFileName);
+        attachment = uploadPictureFile(file,attachment);
 
         attachmentDao.insert(attachment);
 
@@ -73,24 +67,14 @@ public class AttachmentServiceImpl implements AttachmentService {
      */
     public int update(String attachmentId, MultipartFile file) {
         Attachment attachment = findByAttachmentId(attachmentId);
+        if(null == attachment){
+            return 0;
+        }
+        if( true == deleteAttachmentFile( attachment ) ){
+            return attachmentDao.delete(attachmentId);
+        }
 
-        //得到当前的原图文件名
-        String originalImageFileName = attachment.getFilePath().replaceAll(fileRelativePath, "");
-        //得到当前的缩略图文件名
-        String narrowImageFileName = attachment.getNarrowImagePath().replaceAll(fileRelativePath, "");
-        //执行删除操作
-        FileUtil.delFile(fileAbsoluteathPath.replaceAll("WEB-INF/classes/../../static/attachment/", "") + fileRelativePath + originalImageFileName);
-        FileUtil.delFile(fileAbsoluteathPath.replaceAll("WEB-INF/classes/../../static/attachment/", "") + fileRelativePath + narrowImageFileName);
-
-        //执行上传 返回真实文件名
-        originalImageFileName = FileUpload.fileUp(file, fileAbsoluteathPath, UuidUtil.get32UUID());
-        //得到压缩图文件名
-        narrowImageFileName = "_" + originalImageFileName;
-        //进行压缩
-        NarrowImage.imageNarrow(fileAbsoluteathPath, narrowImageFileName, originalImageFileName, 5);
-
-        attachment.setFilePath(fileRelativePath + originalImageFileName);
-        attachment.setNarrowImagePath(fileRelativePath + narrowImageFileName);
+        attachment = uploadPictureFile(file,attachment);
 
         return attachmentDao.update(attachment);
     }
@@ -103,35 +87,63 @@ public class AttachmentServiceImpl implements AttachmentService {
      */
     public int delete(String attachmentId) {
         Attachment attachment = findByAttachmentId(attachmentId);        //得到当前的原图文件名
-        String originalImageFileName = attachment.getFilePath().replaceAll(fileRelativePath, "");
-        //得到当前的缩略图文件名
-        String narrowImageFileName = attachment.getNarrowImagePath().replaceAll(fileRelativePath, "");
-        //执行删除操作
-        FileUtil.delFile(fileAbsoluteathPath.replaceAll("WEB-INF/classes/../../static/attachment/", "") + fileRelativePath + originalImageFileName);
-        FileUtil.delFile(fileAbsoluteathPath.replaceAll("WEB-INF/classes/../../static/attachment/", "") + fileRelativePath + narrowImageFileName);
+        if(null == attachment){
+            return 0;
+        }
+        if( true == deleteAttachmentFile( attachment ) ){
+            return attachmentDao.delete(attachmentId);
+        }
+        return 0;
 
-        return attachmentDao.delete(attachmentId);
     }
 
     /**
-     * 删除一条信息
-     *
+     * 删除
      * @param contentId
      * @return 成功删除的信息条数
      */
     public int deleteByContentId(String contentId){
         System.out.println("service"+contentId);
-        Attachment attachment = attachmentDao.findByContentId(contentId).get(0);        //得到当前的原图文件名
-        if(null == attachment){
+        List<Attachment> attachmentList = attachmentDao.findByContentId(contentId);        //得到当前的原图文件名
+        if(0 == attachmentList.size()){
             return 0;
         }
+        for(int i=0;i<attachmentList.size();i++){
+            deleteAttachmentFile( attachmentList.get(i) );
+        }
+        return attachmentDao.deleteByContentId(contentId);
+
+    }
+    /**
+     * 上传并设置Attachment文件路径和压缩文件路径
+     * @param file
+     * @param attachment
+     * @return
+     */
+    Attachment uploadPictureFile(MultipartFile file,Attachment attachment){
+        //执行上传 返回真实文件名
+        String fileName = FileUpload.fileUp(file, fileAbsoluteathPath, UuidUtil.get32UUID());
+        //得到压缩图文件名
+        String narrowImageFileName = "_" + fileName;
+        //进行压缩
+        NarrowImage.imageNarrow(fileAbsoluteathPath, narrowImageFileName, fileName, 5);
+        attachment.setFilePath(fileRelativePath + fileName);
+        attachment.setNarrowImagePath(fileRelativePath + narrowImageFileName);
+        return attachment;
+    }
+
+    /**
+     * 删除attachment相应的文件
+     * @param attachment
+     * @return
+     */
+    boolean deleteAttachmentFile(Attachment attachment){
         String originalImageFileName = attachment.getFilePath().replaceAll(fileRelativePath, "");
         //得到当前的缩略图文件名
         String narrowImageFileName = attachment.getNarrowImagePath().replaceAll(fileRelativePath, "");
         //执行删除操作
         FileUtil.delFile(fileAbsoluteathPath.replaceAll("WEB-INF/classes/../../static/attachment/", "") + fileRelativePath + originalImageFileName);
         FileUtil.delFile(fileAbsoluteathPath.replaceAll("WEB-INF/classes/../../static/attachment/", "") + fileRelativePath + narrowImageFileName);
-
-        return attachmentDao.deleteByContentId(contentId);
+        return true;
     }
 }
